do: fix compile error
