--- src/main/java/com/posomo/project/core/entity/DateEntity.java
@@ -1,19 +1,0 @@
-package com.posomo.project.core.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public abstract class DateEntity extends BaseEntity{
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-}
