--- src/main/java/com/posomo/project/core/entity/OwnEntity.java
@@ -1,12 +1,0 @@
-package com.posomo.project.core.entity;
-
-
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.LastModifiedBy;
-
-@Getter
-public abstract class OwnEntity extends DateEntity{
-    @CreatedBy
-    private Long ownerId;
-}
