--- src/main/java/com/posomo/project/core/enums/Day.java
@@ -1,5 +1,0 @@
-package com.posomo.project.core.enums;
-
-public enum Day {
-    MON,TUE,WED,THU,FRI,SAT,SUN
-}
