--- src/main/java/com/posomo/project/domain/restaurant/entity/RestaurantTimeType.java
@@ -1,5 +1,0 @@
-package com.posomo.project.domain.restaurant.entity;
-
-public enum RestaurantTimeType {
-    BREAK,OPEN,LAST_ORDER
-}
