--- src/main/java/com/posomo/project/domain/restaurant/entity/common/ManyToOne/RestaurantManyToOneDateEntity.java
@@ -1,21 +1,0 @@
-package com.posomo.project.domain.restaurant.entity.common.ManyToOne;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public class RestaurantManyToOneDateEntity extends RestaurantManyToOneSubEntity {
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
-}
