+++ src/main/java/com/posomo/project/domain/restaurant/entity/core/entity/Category.java
@@ -1,11 +1,10 @@
-package com.posomo.project.core.entity;
+package com.posomo.project.domain.restaurant.entity.core.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import lombok.Getter;
-import lombok.Setter;
 
 @Entity
 @Getter
