+++ src/main/java/com/posomo/project/domain/restaurant/entity/core/enums/Day.java
@@ -1,0 +1,5 @@
+package com.posomo.project.domain.restaurant.entity.core.enums;
+
+public enum Day {
+    MON,TUE,WED,THU,FRI,SAT,SUN
+}
