+++ src/main/java/com/posomo/project/domain/restaurant/entity/manyToOne/RestaurantTimeType.java
@@ -1,0 +1,5 @@
+package com.posomo.project.domain.restaurant.entity.manyToOne;
+
+public enum RestaurantTimeType {
+    BREAK,OPEN,LAST_ORDER
+}
